proj/get_youtube_video_list/requirements.txt
requests==2.22.0
google-api-python-client==1.12.5
google-auth==1.23.0